Rename redwood-schema-generator to redwood-generator
Also remove "schema" part of package name.
