+++ redwood-generator/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-generator
+POM_NAME=Redwood generator
