+++ redwood-generator/src/test/kotlin/app/cash/redwood/generator/DiffConsumingGenerationTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.schema.generator
+package app.cash.redwood.generator
 
 import app.cash.redwood.schema.Property
 import app.cash.redwood.schema.Schema
@@ -21,29 +21,34 @@ import app.cash.redwood.schema.Widget
 import app.cash.redwood.schema.parser.parseSchema
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
+import java.util.regex.Pattern
+import java.util.regex.Pattern.MULTILINE
 
-class SchemaGenerationTest {
+class DiffConsumingGenerationTest {
   @Schema(
     [
-      DataClassWidget::class,
-      ObjectWidget::class,
+      Node12::class,
+      Node1::class,
+      Node3::class,
+      Node2::class,
     ]
   )
-  interface TestSchema
+  interface SortedByTagSchema
   @Widget(1)
-  data class DataClassWidget(
-    @Property(1) val label: String,
-  )
+  data class Node1(@Property(1) val text: String)
   @Widget(2)
-  object ObjectWidget
+  data class Node2(@Property(1) val text: String)
+  @Widget(3)
+  data class Node3(@Property(1) val text: String)
+  @Widget(12)
+  data class Node12(@Property(1) val text: String)
 
-  @Test fun schemaInstanceCreation() {
-    val schema = parseSchema(TestSchema::class)
+  @Test fun `names are sorted by their node tags`() {
+    val schema = parseSchema(SortedByTagSchema::class)
 
-    val dataClassWidget = generateSchemaWidget(schema, schema.widgets.single { it.type == DataClassWidget::class })
-    assertThat(dataClassWidget.toString()).contains("get() = SchemaGenerationTest.DataClassWidget(\n")
-
-    val objectWidget = generateSchemaWidget(schema, schema.widgets.single { it.type == ObjectWidget::class })
-    assertThat(objectWidget.toString()).contains("get() = SchemaGenerationTest.ObjectWidget\n")
+    val fileSpec = generateDiffConsumingWidgetFactory(schema)
+    assertThat(fileSpec.toString()).containsMatch(
+      Pattern.compile("1 ->[^2]+2 ->[^3]+3 ->[^1]+12 ->", MULTILINE)
+    )
   }
 }
