+++ redwood-generator/src/test/kotlin/app/cash/redwood/generator/DiffProducingGenerationTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.schema.generator
+package app.cash.redwood.generator
 
 import app.cash.redwood.schema.Property
 import app.cash.redwood.schema.Schema
@@ -22,28 +22,29 @@ import app.cash.redwood.schema.parser.parseSchema
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
 
-class WidgetGenerationTest {
+class DiffProducingGenerationTest {
   @Schema(
     [
-      NavigationBar.Button::class,
-      Button::class,
+      IdPropertyNameCollisionNode::class,
     ]
   )
-  interface SimpleNameCollisionSchema
-  interface NavigationBar {
-    @Widget(1)
-    data class Button(@Property(1) val text: String)
-  }
-  @Widget(3)
-  data class Button(@Property(1) val text: String)
+  interface IdPropertyNameCollisionSchema
+  @Widget(1)
+  data class IdPropertyNameCollisionNode(
+    @Property(1) val label: String,
+    @Property(2) val id: String,
+  )
 
-  @Test fun `simple names do not collide`() {
-    val schema = parseSchema(SimpleNameCollisionSchema::class)
+  @Test fun `id property does not collide`() {
+    val schema = parseSchema(IdPropertyNameCollisionSchema::class)
 
-    val fileSpec = generateWidgetFactory(schema)
-    assertThat(fileSpec.toString()).apply {
-      contains("fun WidgetGenerationTestNavigationBarButton()")
-      contains("fun WidgetGenerationTestButton()")
-    }
+    val fileSpec = generateDiffProducingWidget(schema, schema.widgets.single())
+    assertThat(fileSpec.toString()).contains(
+      """
+      |  public override fun id(id: String): Unit {
+      |    appendDiff(PropertyDiff(this.id, 2, json.encodeToJsonElement(serializer_0, id)))
+      |  }
+      |""".trimMargin()
+    )
   }
 }
