+++ redwood-generator/src/test/kotlin/app/cash/redwood/generator/WidgetGenerationTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.schema.generator
+package app.cash.redwood.generator
 
 import app.cash.redwood.schema.Property
 import app.cash.redwood.schema.Schema
@@ -21,34 +21,29 @@ import app.cash.redwood.schema.Widget
 import app.cash.redwood.schema.parser.parseSchema
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
-import java.util.regex.Pattern
-import java.util.regex.Pattern.MULTILINE
 
-class DiffConsumingGenerationTest {
+class WidgetGenerationTest {
   @Schema(
     [
-      Node12::class,
-      Node1::class,
-      Node3::class,
-      Node2::class,
+      NavigationBar.Button::class,
+      Button::class,
     ]
   )
-  interface SortedByTagSchema
-  @Widget(1)
-  data class Node1(@Property(1) val text: String)
-  @Widget(2)
-  data class Node2(@Property(1) val text: String)
+  interface SimpleNameCollisionSchema
+  interface NavigationBar {
+    @Widget(1)
+    data class Button(@Property(1) val text: String)
+  }
   @Widget(3)
-  data class Node3(@Property(1) val text: String)
-  @Widget(12)
-  data class Node12(@Property(1) val text: String)
+  data class Button(@Property(1) val text: String)
 
-  @Test fun `names are sorted by their node tags`() {
-    val schema = parseSchema(SortedByTagSchema::class)
+  @Test fun `simple names do not collide`() {
+    val schema = parseSchema(SimpleNameCollisionSchema::class)
 
-    val fileSpec = generateDiffConsumingWidgetFactory(schema)
-    assertThat(fileSpec.toString()).containsMatch(
-      Pattern.compile("1 ->[^2]+2 ->[^3]+3 ->[^1]+12 ->", MULTILINE)
-    )
+    val fileSpec = generateWidgetFactory(schema)
+    assertThat(fileSpec.toString()).apply {
+      contains("fun WidgetGenerationTestNavigationBarButton()")
+      contains("fun WidgetGenerationTestButton()")
+    }
   }
 }
