--- redwood-schema-generator/gradle.properties
@@ -1,2 +1,0 @@
-POM_ARTIFACT_ID=redwood-schema-generator
-POM_NAME=Redwood schema generator
