+++ settings.gradle
@@ -3,13 +3,13 @@ rootProject.name = 'redwood'
 include ':compose:runtime'
 
 include ':redwood-compose'
+include ':redwood-generator'
 include ':redwood-gradle-plugin'
 include ':redwood-protocol'
 include ':redwood-protocol-compose'
 include ':redwood-protocol-widget'
 include ':redwood-schema'
 include ':redwood-schema-annotations'
-include ':redwood-schema-generator'
 include ':redwood-widget'
 
 include ':test-schema'
