+++ test-schema/compose/build.gradle
@@ -42,13 +42,13 @@ configurations {
 }
 
 dependencies {
-  generator projects.redwoodSchemaGenerator
+  generator projects.redwoodGenerator
   generator projects.testSchema
 }
 
 def generate = tasks.register('redwoodGenerate', JavaExec) {
   it.classpath(configurations.generator)
-  it.main('app.cash.redwood.schema.generator.Main')
+  it.main('app.cash.redwood.generator.Main')
   it.args(
     '--compose',
     '--out', generatedDir,
