+++ test-schema/compose/protocol/build.gradle
@@ -41,13 +41,13 @@ configurations {
 }
 
 dependencies {
-  generator projects.redwoodSchemaGenerator
+  generator projects.redwoodGenerator
   generator projects.testSchema
 }
 
 def generate = tasks.register('redwoodGenerate', JavaExec) {
   it.classpath(configurations.generator)
-  it.main('app.cash.redwood.schema.generator.Main')
+  it.main('app.cash.redwood.generator.Main')
   it.args(
     '--compose-protocol',
     '--out', generatedDir,
