+++ test-schema/widget/build.gradle
@@ -40,13 +40,13 @@ configurations {
 }
 
 dependencies {
-  generator projects.redwoodSchemaGenerator
+  generator projects.redwoodGenerator
   generator projects.testSchema
 }
 
 def generate = tasks.register('redwoodGenerate', JavaExec) {
   it.classpath(configurations.generator)
-  it.main('app.cash.redwood.schema.generator.Main')
+  it.main('app.cash.redwood.generator.Main')
   it.args(
     '--widget',
     '--out', generatedDir,
